Stage one of Simple Shell
Working as it should, prompts user for input with an indicator, for now it is simply "> " in future plan to change this to show a working directory before. Takes users string input, separates each word by the spaces around them and displays each token in the form of different lines and inside of quotation works, example "ls" "-w".